Add KGIIR Architectural Record (1.11837 BPB)
